+++ src/abi/index.ts
@@ -1,4 +1,5 @@
 export * from './authorizer';
+export * from './balancerBatchRouter';
 export * from './balancerQueries';
 export * from './balancerBatchRouter';
 export * from './balancerRelayer';
@@ -10,7 +11,8 @@ export * from './erc20';
 export * from './permit2';
 export * from './vaultV2';
 export * from './vaultV3';
+export * from './vaultExtensionV3';
 export * from './weightedPoolFactory.V3';
 export * from './weightedPoolFactoryV4.V2';
 export * from './weightedPoolV4.V2';
-export * from './vaultExtensionV3';
+export * from './weightedPool.V3';