+++ src/data/providers/balancer-api/index.ts
@@ -4,7 +4,11 @@ import { ChainId } from '../../../utils';
 import { NestedPools } from './modules/nested-pool-state';
 import { SorSwapPaths } from './modules/sorSwapPaths';
 
-export { mapPoolToNestedPoolState } from './modules/nested-pool-state';
+export { SorInput as GetQuoteInput } from './modules/sorSwapPaths';
+export {
+    mapPoolToNestedPoolState,
+    PoolGetPool,
+} from './modules/nested-pool-state';
 
 export class BalancerApi {
     balancerApiClient: BalancerApiClient;