+++ src/entities/swap/swaps/v2/index.ts
@@ -125,6 +125,7 @@ export class SwapV2 implements SwapBase {
                     this.outputAmount.token,
                     result,
                 ),
+                amountIn: this.inputAmount,
             };
         }
         return {
@@ -133,6 +134,7 @@ export class SwapV2 implements SwapBase {
                 this.inputAmount.token,
                 result,
             ),
+            amountOut: this.outputAmount,
         };
     }
 
@@ -163,6 +165,7 @@ export class SwapV2 implements SwapBase {
                         ],
                     ),
                 ),
+                amountIn: this.inputAmount,
             };
         }
         return {
@@ -173,6 +176,7 @@ export class SwapV2 implements SwapBase {
                     result[this.assets.indexOf(this.inputAmount.token.address)],
                 ),
             ),
+            amountOut: this.outputAmount,
         };
     }
 
@@ -287,18 +291,22 @@ export class SwapV2 implements SwapBase {
         }
         if (this.swapKind === SwapKind.GivenIn) {
             return {
-                to: this.to(),
+                to: VAULT[this.chainId],
                 callData,
                 value: this.value(limitAmount, !!input.wethIsEth),
                 minAmountOut: limitAmount,
             };
         }
         return {
-            to: this.to(),
+            to: VAULT[this.chainId],
             callData,
             value: this.value(limitAmount, !!input.wethIsEth),
             maxAmountIn: limitAmount,
         };
+    }
+
+    buildCallWithPermit2(): SwapBuildOutputExactIn | SwapBuildOutputExactOut {
+        throw new Error('buildCallWithPermit2 is not supported on v2');
     }
 
     /**
@@ -319,10 +327,6 @@ export class SwapV2 implements SwapBase {
             else value = limitAmount.amount;
         }
         return value;
-    }
-
-    private to(): Address {
-        return VAULT[this.chainId];
     }
 
     /**