+++ src/entities/utils/calculateProportionalAmounts.ts
@@ -1,12 +1,18 @@
 import { Address, parseUnits } from 'viem';
 import { InputAmount } from '@/types';
 import { HumanAmount } from '@/data';
+import { MathSol } from '@/utils';
 
 /**
  * For a given pool, calculate all token amounts proportional to a given reference amount.
+ *
+ * Note: when using this helper to build an AddLiquidityProportional input,
+ * please mind that referenceAmount should be relative to the token that the user
+ * has the lowest balance compared to the pool's proportions. Otherwise the transaction
+ * may require more balance than the user has.
  * @param pool
  * @param referenceAmount
- * @returns
+ * @returns Proportional amounts rounded down to the nearest integer
  */
 export function calculateProportionalAmounts(
     pool: {
@@ -50,8 +56,11 @@ export function calculateProportionalAmounts(
 
     // calculate proportional amounts
     const referenceTokenBalance = balances[referenceTokenIndex];
-    const proportionalAmounts = balances.map(
-        (b) => (b * referenceAmount.rawAmount) / referenceTokenBalance,
+    const proportionalAmounts = balances.map((b) =>
+        MathSol.divDownFixed(
+            MathSol.mulDownFixed(b, referenceAmount.rawAmount),
+            referenceTokenBalance,
+        ),
     );
 
     const amounts = tokensWithBpt.map(({ address, decimals }, index) => ({